refactor(backup): consolidate backup filename constants and update imports
