+++ src/app/imex/file-imex/file-imex.component.ts
@@ -10,7 +10,11 @@ import { HttpClient } from '@angular/common/http';
 import { MatDialog } from '@angular/material/dialog';
 import { SnackService } from '../../core/snack/snack.service';
 import { download } from '../../util/download';
-import { getBackupTimestamp } from '../../util/get-backup-timestamp';
+import {
+  BACKUP_FILENAME_PREFIX,
+  BACKUP_FILENAME_PREFIX_ANONYMIZED,
+  getBackupTimestamp,
+} from '../../util/get-backup-timestamp';
 import { DialogImportFromUrlComponent } from '../dialog-import-from-url/dialog-import-from-url.component';
 import { T } from '../../t.const';
 import { TODAY_TAG } from '../../features/tag/tag.const';
@@ -261,7 +265,7 @@ export class FileImexComponent implements OnInit {
 
   async downloadBackup(): Promise<void> {
     const data = await this._backupService.loadCompleteBackup(true);
-    const fileName = `sp-backup_${getBackupTimestamp()}.json`;
+    const fileName = `${BACKUP_FILENAME_PREFIX}_${getBackupTimestamp()}.json`;
     const result = await download(fileName, JSON.stringify(data));
     if ((IS_NATIVE_PLATFORM && !result.wasCanceled) || result.isSnap) {
       this._snackService.open({
@@ -275,7 +279,7 @@ export class FileImexComponent implements OnInit {
 
   async privacyAppDataDownload(): Promise<void> {
     const data = await this._backupService.loadCompleteBackup(true);
-    const fileName = `sp-backup-anonymized_${getBackupTimestamp()}.json`;
+    const fileName = `${BACKUP_FILENAME_PREFIX_ANONYMIZED}_${getBackupTimestamp()}.json`;
     const result = await download(fileName, privacyExport(data));
     if ((IS_NATIVE_PLATFORM && !result.wasCanceled) || result.isSnap) {
       this._snackService.open({
